+++ Kokoro/APIv2/OnDiscard.cs
@@ -35,6 +35,16 @@ public partial interface IKokoroApi
 				/// The actual action to run on discard.
 				/// </summary>
 				CardAction Action { get; set; }
+				
+				/// <summary>
+				/// Whether to show the icon for the wrapper action.
+				/// </summary>
+				bool ShowOnDiscardIcon { get; set; }
+				
+				/// <summary>
+				/// Whether to show the tooltip for the wrapper action.
+				/// </summary>
+				bool ShowOnDiscardTooltip { get; set; }
 
 				/// <summary>
 				/// Sets <see cref="Action"/>.
@@ -42,6 +52,20 @@ public partial interface IKokoroApi
 				/// <param name="value">The new value.</param>
 				/// <returns>This object after the change.</returns>
 				IOnDiscardAction SetAction(CardAction value);
+
+				/// <summary>
+				/// Sets <see cref="ShowOnDiscardIcon"/>.
+				/// </summary>
+				/// <param name="value">The new value.</param>
+				/// <returns>This object after the change.</returns>
+				IOnDiscardAction SetShowOnDiscardIcon(bool value);
+
+				/// <summary>
+				/// Sets <see cref="ShowOnDiscardTooltip"/>.
+				/// </summary>
+				/// <param name="value">The new value.</param>
+				/// <returns>This object after the change.</returns>
+				IOnDiscardAction SetShowOnDiscardTooltip(bool value);
 			}
 		}
 	}